requirements.txt
asgiref==3.8.1
autoflake==2.3.1
autopep8==2.3.2
black==25.1.0
certifi==2025.1.31
cfgv==3.4.0
charset-normalizer==3.4.1
click==8.1.8
distlib==0.3.9
dj-rest-auth==7.0.1
Django==5.1.6
django-allauth==65.4.1
django-ckeditor-5==0.2.17
django-cors-headers==4.7.0
django-extensions==3.2.3
django-filter==25.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-nested-routers==0.94.1
drf-yasg==1.21.8
filelock==3.17.0
identify==2.6.7
idna==3.10
inflection==0.5.1
mypy-extensions==1.0.0
nodeenv==1.9.1
packaging==24.2
pathspec==0.12.1
pillow==11.1.0
platformdirs==4.3.6
pre_commit==4.1.0
psycopg2-binary==2.9.10
pycodestyle==2.12.1
pyflakes==3.2.0
PyJWT==2.10.1
python-dotenv==1.0.1
pytz==2025.1
pyupgrade==3.19.1
PyYAML==6.0.2
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
tokenize_rt==6.1.0
tomli==2.2.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
virtualenv==20.29.2
